+++ dev.py
@@ -1,5 +1,5 @@
 from flask import Flask, render_template
-from models.db import db
+from instance.db import db
 from src.authentication import auth
 from src.user_profile import profile
 from src.search import search
