+++ models/catalogue/book.py
