+++ models/catalogue/catalogue_manager.py
