+++ src/book.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, render_template, request, redirect, url_for, flash, session
-from src.db_manager import DBManager
+from models.database_manager.db_manager import DBManager
 
 book = Blueprint('book', __name__)
